requirements.txt
blinker==1.4
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.6.0
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
numpy==2.2.6
oauthlib==3.2.0
pandas==2.3.3
platformdirs==2.5.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.4.1
safe-pysha3==1.0.5
SecretStorage==3.3.1
six==1.16.0
tqdm==4.67.3
tzdata==2025.3
unattended-upgrades==0.1
virtualenv==20.13.0+ds
wadllib==1.3.6
zipp==1.0.0
